+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Fixed
 
+- Project specific CD user not permissioned on provision of new component repository ([#670](https://github.com/opendevstack/ods-provisioning-app/issues/670))
 - Show an error message when the component id does not match the regex expression ([#624](https://github.com/opendevstack/ods-provisioning-app/issues/624))
 - Set confluence shortcut in jira even if confluence server is not activated ([#613](https://github.com/opendevstack/ods-provisioning-app/issues/613))
 - Missing verification of confluence project already exists on project creation event ([#607](https://github.com/opendevstack/ods-provisioning-app/issues/607))